% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCalibrationPlot.R
\name{generateCalibrationPlot}
\alias{generateCalibrationPlot}
\title{generateCalibrationPlot}
\usage{
generateCalibrationPlot(
  validationResults,
  smoothMethod = "loess",
  span = 0.75,
  se = T,
  scatter = F,
  hist = T,
  hist_norm = 0.2,
  subgroup = "all",
  outputFile = "results/plot/calibration.pdf"
)
}
\arguments{
\item{validationResults}{The allValidationResults object generated by the generateSubgroupValidationData function.}

\item{smoothMethod}{The smooth method when plotting.}

\item{span}{Plot span value.}

\item{se}{If standard error included.}

\item{scatter}{If scatter included.}

\item{hist}{If histogram included.}

\item{hist_norm}{Normalization value for histogram.}

\item{subgroup}{The name of the subgroup. For a binary subgroup, the name is followed by #1 or #0 to determine the subgroup.}

\item{outputFile}{The output file path including the file name, such as plot/AUPRC_forest_plot.tiff.}
}
\value{
It saves and returns the forest plot as a tiff file.
}
\description{
Generate calibration plot for the given subgroup population.
}
