% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateForestPlot.R
\name{generateForestPlot}
\alias{generateForestPlot}
\title{generateForestPlot}
\usage{
generateForestPlot(
  csvFile = NULL,
  resultDataFrame = NULL,
  estimateCol = "AUROC",
  outputFile = "AUPRC_forest_plot.tiff"
)
}
\arguments{
\item{csvFile}{If the subgroup validation table generated by the generateSubgroupValidationData function has been saved in a CSV file, it can be used as input here.}

\item{resultDataFrame}{The subgroup validation table generated by the generateSubgroupValidationData function.}

\item{estimateCol}{It can be set to AUROC or AUPRC to plot the forest diagram accordingly.}

\item{outputFile}{The output file path including the file name, such as plot/AUPRC_forest_plot.tiff}
}
\value{
It saves and returns the forest plot as a tiff file.
}
\description{
Generate forest plots based on AUROC and AUPRC values
}
