% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateForestPlot.R
\name{generateMultiForestPlot}
\alias{generateMultiForestPlot}
\title{generateMultiForestPlot}
\usage{
generateMultiForestPlot(
  csvFile = NULL,
  resultDataFrame = NULL,
  outputFile = "AUPRC_forest_plot.tiff"
)
}
\arguments{
\item{csvFile}{If the subgroup validation table generated by the generateSubgroupValidationData function has been saved in a CSV file, it can be used as input here.}

\item{resultDataFrame}{The subgroup validation table generated by the generateSubgroupValidationData function.}

\item{outputFile}{The output file path including the file name, such as plot/AUPRC_forest_plot.tiff}
}
\value{
It saves and returns the forest plot as a tiff file.
}
\description{
Generate forest plots for AUROC and AUPRC values in a single plot
}
