% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNetBenefitPlot.R
\name{generateNetBenefitPlot}
\alias{generateNetBenefitPlot}
\title{generateNetBenefitPlot}
\usage{
generateNetBenefitPlot(plpPrediction, subgroup, outputFile)
}
\arguments{
\item{plpPrediction}{The allValidationResults object generated by the generateSubgroupValidationData function.}

\item{subgroup}{The name of the subgroup. For a binary subgroup, the name is followed by #1 or #0 to determine the subgroup.}

\item{outputFile}{The output file path including the file name, such as plot/netbenefit_plot.pdf}
}
\value{
It saves and returns the forest plot as a pdf file.
}
\description{
Generate standardized net benefit plot for the given subgroup population.
}
