% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSubgroupValidationData.R
\name{generateSubgroupValidationData}
\alias{generateSubgroupValidationData}
\title{generateSubgroupValidationData}
\usage{
generateSubgroupValidationData(
  connectionDetails,
  plpValidationRds,
  evaluationType,
  subgroupCohortInfo
)
}
\arguments{
\item{connectionDetails}{connectionDetails object to be connected to the CDM database.}

\item{plpValidationRds}{the rds file associated with the PLP model that has been run on the whole target cohort.}

\item{evaluationType}{It can be one of the following items: Test or weakRecalibration.}
}
\value{
a list of three tables including PLP subgroup validation table, subgroup performance metrics table, subgroup calibration summary table.
}
\description{
Generate cohort subgroup validation data for the already run PLP model
}
